requirements.txt
# =============================================================================
# REQUIREMENTS.TXT - SMARTETUDE
# =============================================================================
# Dépendances Python pour SmartEtude
# 
# Installation : pip install -r requirements.txt
# Mise à jour : pip install -r requirements.txt --upgrade
# 
# Auteur: [Votre Nom]
# Date: 2024
# Version: 1.0.0

# =============================================================================
# DJANGO & CORE FRAMEWORK
# =============================================================================

# Framework web principal
Django==4.2.7                    # Version LTS stable
djangorestframework==3.14.0      # API REST professionnelle
django-cors-headers==4.3.1       # Gestion CORS
django-filter==23.3              # Filtrage avancé
django-extensions==3.2.3         # Extensions utiles

# =============================================================================
# BASE DE DONNÉES & CACHE
# =============================================================================

# PostgreSQL (production)
psycopg2-binary==2.9.7          # Adaptateur PostgreSQL
# psycopg2==2.9.7               # Alternative sans binaire

# Redis et cache
redis==5.0.1                     # Client Redis
django-redis==5.4.0              # Intégration Django-Redis

# =============================================================================
# INTELLIGENCE ARTIFICIELLE & MACHINE LEARNING
# =============================================================================

# Note: l'application utilise désormais l'API OpenAI (HTTP) et n'exécute plus
# de modèles locaux. Les dépendances lourdes suivantes ont été retirées:
# - transformers, tokenizers, torch, accelerate, sentencepiece
# Conservez uniquement 'requests' (voir section utilitaires) pour les appels API.

# Traitement des données
numpy==1.24.3                    # Calculs numériques
pandas==2.0.3                    # Manipulation de données
scikit-learn==1.3.2              # Algorithmes ML classiques

# Traitement du langage naturel (optionnel)
# nltk==3.8.1                      # Natural Language Toolkit
# spacy==3.7.2                     # NLP industriel

# =============================================================================
# TRAITEMENT DE DOCUMENTS & PDF
# =============================================================================

# Extraction de texte
PyPDF2==3.0.1                    # Lecture de PDF
python-docx==1.1.0               # Lecture de DOCX
Pillow==10.1.0                   # Traitement d'images
pytesseract==0.3.10              # OCR (reconnaissance de texte)

# =============================================================================
# TÂCHES ASYNCHRONES & BACKGROUND
# =============================================================================

# Celery pour le traitement asynchrone
celery==5.3.4                    # Queue de tâches
django-celery-beat==2.5.0        # Planification de tâches
django-celery-results==2.5.1     # Stockage des résultats

# =============================================================================
# SÉCURITÉ & AUTHENTIFICATION
# =============================================================================

# JWT et OAuth
djangorestframework-simplejwt==5.3.0  # Authentification JWT
django-oauth-toolkit==2.3.0           # OAuth2 provider
django-ratelimit==4.1.0               # Limitation de débit

# =============================================================================
# API & SÉRIALISATION
# =============================================================================

# Documentation automatique
drf-spectacular==0.26.5          # OpenAPI/Swagger moderne
drf-yasg==1.21.7                 # Alternative Swagger

# =============================================================================
# MONITORING & LOGGING
# =============================================================================

# Sentry pour le monitoring d'erreurs
sentry-sdk==1.38.0               # SDK Sentry officiel

# Debug et développement
django-debug-toolbar==4.2.0      # Barre de debug Django

# =============================================================================
# TESTS & QUALITÉ DE CODE
# =============================================================================

# Framework de tests
pytest==7.4.3                    # Framework de tests moderne
pytest-django==4.7.0             # Intégration Django-Pytest
coverage==7.3.2                   # Couverture de code

# Formatage et linting
black==23.11.0                    # Formateur de code
flake8==6.1.0                     # Linter Python
isort==5.12.0                     # Tri des imports

# =============================================================================
# DÉPLOIEMENT & PRODUCTION
# =============================================================================

# Serveur WSGI
gunicorn==21.2.0                 # Serveur WSGI production
whitenoise==6.6.0                 # Fichiers statiques

# Configuration d'environnement
python-decouple==3.8              # Gestion des variables d'env

# =============================================================================
# UTILITAIRES & BIBLIOTHÈQUES
# =============================================================================

# Manipulation de dates
python-dateutil==2.8.2            # Utilitaires de dates
packaging>=23.1,<25               # Outil de gestion de versions (utilisé par l'app)

# Requêtes HTTP
requests==2.31.0                  # Client HTTP
urllib3==2.0.7                    # HTTP client (dépendance)

# =============================================================================
# DÉPENDANCES OPTIONNELLES (décommentez selon vos besoins)
# =============================================================================

# # Base de données alternatives
# mysqlclient==2.2.0              # MySQL (si nécessaire)
# cx-Oracle==8.3.0                # Oracle (si nécessaire)

# # Cache alternatif
# django-cacheops==8.0.0          # Cache avancé avec ORM

# # Sécurité renforcée
# django-axes==6.1.0              # Protection contre les attaques
# django-defender==0.9.0          # Blocage d'IP malveillantes

# # Performance et optimisation
# django-query-profiler==0.1.0    # Profilage des requêtes
# django-silk==5.0.0              # Profiling avancé

# # Monitoring avancé
# prometheus-client==0.19.0        # Métriques Prometheus
# django-prometheus==2.3.1         # Métriques Django

# # Développement local
# django-extensions==3.2.3         # Extensions de développement
# ipython==8.17.2                  # Shell Python amélioré
# werkzeug==3.0.1                  # Serveur de développement
